separated session function to its own file and eliminated loacal variables count that was disabling it
