+++ tickers/tickers_BTC.txt
@@ -1,271 +1,228 @@
-COTI
-GRS
-REEF
-GTC
-WTC
-POND
-FTT
-RCN
-BZRX
-DUSK
-POLY
-QKC
-DIA
-STRAX
-CELR
-SAND
-NAV
-AUTO
-DEGO
-CKB
-YOYO
-RIF
-BADGER
-ONE
-GXS
-BEAM
-FRONT
-NEAR
-AUCTION
-IDEX
-POWR
-GRT
-EZ
-OMG
-JST
-TOMO
-SFP
-CHR
-RLC
-WAN
-SYS
-ZEN
-NANO
-PHB
-GVT
-KAVA
-TFUEL
-VIDT
-DOGE
-ALICE
-POLS
-UNI
-GLM
-EPS
-DGB
-UMA
-LSK
-HIVE
-ADA
-CAKE
-QTUM
-CHZ
-QLC
-DATA
-ARDR
-WAVES
-ATA
-XMR
-TKO
-ZIL
-ANT
-FIS
-OXT
-RSR
-MTH
-VIA
-REQ
-KEEP
-LTO
-EGLD
-BEL
-ENJ
-GTO
-VET
-SRM
-AAVE
-FET
-IOTX
-OAX
-ATOM
-ACM
-MIR
-WPR
-DOCK
-SKL
-QSP
-TWT
-ETC
-RVN
-KSM
-AION
-EOS
-AGIX
-REP
-ONT
-AERGO
-CTSI
-LOOM
-FIO
-BAL
-IOTA
-EVX
-ALPHA
-REN
-BTS
-ADX
-RDN
-PIVX
-MANA
-BAT
-ALGO
-BNT
-LPT
-MDT
-ICX
-SKY
-BNB
-ZEC
-CVC
-AXS
-SNX
-SXP
-OGN
-HBAR
-NKN
-NBS
-SC
-NU
-UNFI
-NMR
-FORTH
-IOST
-BAR
-RUNE
-COS
-TRB
-POA
-AVA
-LUNA
-FXS
-XEM
-ELF
-TCT
-ORN
-STEEM
-AKRO
-XLM
-SUSD
-BRD
 NEO
-DODO
-CDT
-WING
-PPT
-KMD
-XVG
-DOT
-GAS
-OG
-DASH
-YFI
-JUV
-TRX
-ARPA
-YFII
-CTK
-THETA
-OM
-BTG
-FUN
-UTK
-FTM
-LINA
-PERP
-MATIC
-WABI
-AVAX
-PAXG
-BCH
-CTXC
-ATM
-AUDIO
-STX
-AST
-INJ
-SNGLS
-SOL
-BAND
-FIL
-RAMP
-FLM
-ICP
-COMP
-AR
-WNXM
-ONG
-ASR
+LRC
+POA
+YOYO
+TROY
 STORJ
-LTC
-SUSHI
-IRIS
-XVS
-ROSE
-ETH
-BCD
-NXS
-LIT
-BQX
-MDX
-SNT
-PSG
-FOR
-DLT
-SNM
-KNC
-CELO
-TORN
-CRV
+NAS
+EVX
+KMD
+DIA
+AERGO
 XRP
-NEBL
-OCEAN
-ARK
-WRX
-MTL
-TLM
-SCRT
-BAKE
-VIB
+SOL
+BAL
+FOR
+NEAR
+BZRX
+REP
+TVK
+BEL
 DNT
 STMX
-ZRX
-XTZ
-NAS
-MKR
-DREP
-LINK
-MITH
-HARD
-PERL
-CND
-PHA
-CFX
-GO
-APPC
-HNT
-FIRO
-TROY
-DCR
-LRC
-MDA
-STPT
-BLZ
-NULS
-TRU
-TVK
-OST
-1INCH
-AMB
+SCRT
 ANKR
-VITE
+CFX
+FTT
+DREP
+BADGER
+CELO
+WAN
+IRIS
+NULS
+XVS
+ADA
+OAX
+BLZ
+TORN
+ROSE
+VIA
+RAMP
+NAV
+POWR
+AR
+OMG
+UMA
+ALICE
+TOMO
+WPR
+ST
+GO
+QKC
+ATOM
+MIR
+WRX
 PNT
+WNXM
+BTS
+W
+FET
+GTC
+XMR
+TRX
+SKY
+CAKE
+SAND
+NANO
+GRT
+RIF
+RDN
+YFII
+RCN
+CELR
+EZ
+FIS
+PHB
+LPT
+UNI
+DGB
+CHR
+HARD
+SYS
+LSK
+ASR
+WAVES
+AAVE
+WING
+QSP
+COS
+AUCTION
+POND
+STPT
+CKB
+SRM
+RSR
+1INCH
+TCT
+CHZ
+OXT
+MANA
+ZEN
+ONE
+SC
+DOCK
+CND
+SNX
+RLC
+DOGE
+BNT
+SFP
+HIVE
+ZIL
+ETC
+TFUEL
+OG
+KLAY
+NKN
+EOS
+ENJ
+QTUM
+STEEM
+ICX
+AST
+BNB
+DCR
+PERP
+NEBL
+SKL
+ALGO
+CTK
+WABI
+AGIX
+VET
+PIVX
+LOOM
+MDA
+POLS
+FIRO
+KAVA
+TRB
+CTSI
+TKO
+DODO
+DATA
+KSM
+GRS
+DUSK
+GAS
+BRD
+CTXC
+GLM
+UNFI
+FORTH
+XLM
+ETH
+ATA
+TWT
+APPC
+IOTX
+ACM
+REEF
+IOTA
+RVN
+BAT
+ALPHA
+EGLD
+AION
+HBAR
+UTK
+MKR
+JST
+LTO
+ELF
+INJ
+CDT
+REQ
+QLC
+PPT
+NXS
+NU
+LUNA
+ORN
+BAND
+TLM
+IOST
+AVA
+RUNE
+ARK
+ONG
+MDT
+FIL
+SNM
+FLM
+DOT
+STX
+PHA
+JUV
+SNT
+NMR
+STRAX
+BQX
+CRV
+COTI
+AXS
+PAXG
+XVG
+BAKE
+MDX
+BEAM
+XEM
+SUSD
+COMP
+FUN
+HNT
+VITE
+DLT
+BAR
+OST
+OCEAN
+FXS
+TRU
+ATM
+MTH
+SUSHI
+YFI
+XTZ
+GVT
+PSG
+REN
